Refactor create_partitions function to support multiple file types and partition dates
