+++ pipelines/datalake/extract_load/sih_gdrive/flows.py
@@ -20,7 +20,7 @@ from pipelines.datalake.extract_load.sih_gdrive.tasks import (
     dowload_from_gdrive,
     transform_data,
 )
-from pipelines.datalake.extract_load.vitacare_api.tasks import create_partitions
+from pipelines.utils.tasks import create_partitions
 from pipelines.datalake.utils.tasks import rename_current_flow_run
 from pipelines.utils.tasks import create_folders, upload_to_datalake
